+++ dca/tl_settings.php
@@ -11,7 +11,7 @@
 
 $tl_settings = &$GLOBALS['TL_DCA']['tl_settings'];
 
-$tl_settings['palettes']['default'] .= ';{wrapper-tags_legend:hide},wrapperTagsAllowedTags,wrapperTagsUseColors';
+$tl_settings['palettes']['default'] .= ';{wrapper-tags_legend:hide},wrapperTagsAllowedTags,wrapperTagsUseColors,wrapperTagsHideValidationStatus';
 
 $tl_settings['fields']['wrapperTagsAllowedTags'] = array
 (
@@ -23,6 +23,13 @@ $tl_settings['fields']['wrapperTagsAllowedTags'] = array
 $tl_settings['fields']['wrapperTagsUseColors'] = array
 (
     'label' => &$GLOBALS['TL_LANG']['tl_settings']['wrapperTagsUseColors'],
+    'inputType' => 'checkbox',
+    'eval' => array('tl_class' => 'w50')
+);
+
+$tl_settings['fields']['wrapperTagsHideValidationStatus'] = array
+(
+    'label' => &$GLOBALS['TL_LANG']['tl_settings']['wrapperTagsHideValidationStatus'],
     'inputType' => 'checkbox',
     'eval' => array('tl_class' => 'w50')
 );